refactor: split iron proxy resource builders
